Correct deprecated require("lspconfig")[...]
`require("lspconfig")[...]` was deprecated in favour of
`vim.lsp.config(...)`, so this updates all of those calls in the Mason
plugin configuration file. Also, `rust-tools` has been abandoned and
users are encourage to migrate to `rustaceanvim`, which is also done in
this commit, since `rust-tools` contained some deprecated calls related
to LSP configuration, as well.